Cleanup the shutdown of the game
In an attempt to fix a random bug where the window sometimes stick around, be
sure to destroy all of the vulkan objects and the xcb window on quit.

The vulkan destruction is rather ugly and hacky. I'm not planning to do a
template RAII type destructor for any of it, but eventually a lot of the
responsibility will be moved to the VulkanDevice as it takes on the
responsibility of creating and allocating objects smarter. Not for garbage
collection, but in order to reuse things like samplers, imageviews, shader
modules, and all sub-allocations through larger buffers