Big change in how descriptor sets are handled.
Not remotely finished, and contains a few warnings from validation layers and a
validation error on quit, but we'll be getting there.

The idea is to split up the pipeline resources into binding frequency. E.g.
things bound by pipeline, such as the camera transforms and projection matrices,
go into a different descriptor layout from the resources needed by materials,
such as material UBOs and textures